chore: simplify DeferredStream to Deferred
